feat: support environment templates
Environment Templates are a part of the spec that allows an Environment
to be defined outside of a Job Template, but still be applied when
running Job(s) created from the Job Template. The CLI doesn't support
these yet.

This commit teaches the CLI about environment templates:
1. In the 'check' command, to verify a template syntax;
2. In the 'schema' command, to generate json schema for
   EnvironmentTemplates; and
3. In the 'run' command, to apply Environments when running a Job
   locally.

The implementation of this also applies a few small cleanups and
refactors to make the implementation more straightforward.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>