feat: Allow job parameters as JSON string
Problem:

We currently allow job parameters to be provided as either KEY=VALUE
pairs or a json/yaml file. However, there are use-cases where one might
have a JSON string but not want to write it to a temp file just to pass
it to the CLI.

Solution:

I've added the option to provide a JSON string as the argument for a job
parameter as well.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>