fix: run subcommand now exits all entered environments
Problem:
 The run subcommand has an incorrect implementation of the Session semantics w.r.t.
which environments should be exited on an action failure. It does not run any action
after an action failure, but it should always be running the onExit for all environments
that were even attempted to be entered.

Solution:
 Rework the LocalSession implementation to split the EnterEnvironment actions out from the
generic action_queue so that we can more easily keep track of which environments were entered.
Then, always exit all environments that were entered and attempted to be entered on an action
failure.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>