feat: Make 'openjd run <template>' help output describe the job parameters
Because a job template is runnable code similar to a script, we can
think of `openjd run <templatefile>` the same as `bash <scriptfile>`.
Part of this is printing help output of the options available.

This change makes it so `openjd run <templatefile> -h` prints help that
is specific about the template file, displaying the job name,
description, and all its parameters. Similarly, if there are missing
required parameters, the error message says that and then displays this
same help text.

Signed-off-by: Mark <399551+mwiebe@users.noreply.github.com>